chore: move the legacy codes into the "legacy" dir and the flatten the packages into pkg
