+++ legacy/controller/risingwave/hook/lifecycle.go
