+++ legacy/controller/risingwave/hook/lifecycle_test.go
