+++ legacy/go.mod
@@ -1,0 +1,1 @@
+// Exclude this directory.
