+++ legacy/manager/compactor.go
