+++ legacy/manager/compute.go
