+++ legacy/manager/consts.go
