+++ legacy/manager/minio.go
