+++ legacy/manager/types.go
