+++ pkg/factory/risingwave_object_factory.go
@@ -30,7 +30,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	risingwavev1alpha1 "github.com/singularity-data/risingwave-operator/apis/risingwave/v1alpha1"
-	"github.com/singularity-data/risingwave-operator/pkg/controllers/risingwave/consts"
+	"github.com/singularity-data/risingwave-operator/pkg/consts"
 )
 
 const (
